Implemented HFD_B2 potential for helium, Quantum cross sections appear to be correctly implemented now.
Cross sections computed for He4-He4, He4-He3, and He3-He3 reproduce those reported by Meeks et al. J. Chem. Phys. (1994),
https://doi.org/10.1063/1.466370

This serves as a test that Bosons, Fermions and distinguishable particles are treated correctly.

+++ cpp/Quantum.cpp
@@ -33,7 +33,7 @@ double bessel_deriv(int n, double r, size_t kind){
 }
 
 Quantum::Quantum(std::string comps) : Spherical(comps, true) {
-    half_spin = std::vector<size_t>(Ncomps, 0);
+    half_spin = std::vector<unsigned int>(Ncomps, 0);
     sigma = vector2d(Ncomps, vector1d(Ncomps, 0.));
     eps = vector2d(Ncomps, vector1d(Ncomps, 0.));
     for (size_t i = 0; i < Ncomps; i++){
